--- src/api/apis/userpost.js
@@ -1,10 +1,0 @@
-import authInstance from '../instance/authInstance';
-
-export const userpostAPI = async (accountname) => {
-  try {
-    const res = await authInstance.get(`/post/${accountname}/userpost`);
-    return res;
-  } catch (err) {
-    console.log(err);
-  }
-};
